Add back links to the supplier assessment pages
This has the side effect of adding a back link to the action plan
session scheduling page, which is not a bad thing.

The back link on the supplier assessment scheduling page is
context-dependent — it either goes back to the intervention progress
page or the view appointment page, depending on whether there’s an
existing appointment. The code for handling this is a bit icky but
I think tolerable.