Display time-sensitive broadcast message
currently used to display a message when delius is unavailable.

We've opted to configure this in its own config file to make it easier
to change this in case we need to display a message like this in future.
It should be fairly easy to extend, if we need to add multiple messages.