Allow custom bearer authentication in RestClient.
This change removes the default authorized client registration from the webclient configuration
allowing each instantiation of a webclient to set its default registration. This is then applied
on the fly in the RestClient for each request. An optional parameter is introduced to the
RestClient methods which allows overriding this value on a per request basis.

The motivation for all this is to allow passing on user scoped to APIs (specifically the
assess-risks-and-needs API which is making this is a requirement).