Make assignment journey use session-based drafts
In the controller tests, I’ve decided to directly mock out the
DraftsService class. I think that in general we’ve only been using
mocking for classes that communicate with an external service, so this
represents a bit of a break from that. But else I’d have had to figure
out how to inject / interrogate session data in those tests, and also
I’d have had to start mocking out the uuid dependency of DraftsService
to get predictable values.

The "backwardsCompability" methods exist in order to make sure that all
of the old routes continue to work. Otherwise we run the risk of causing
an error for somebody who’s halfway through the assignment journey when
we deploy these changes. We can (and should) remove these methods and
routes maybe a week after we deploy.