Make assignment journey use Redis-based drafts
The "backwardsCompability" methods exist in order to make sure that all
of the old routes continue to work. Otherwise we run the risk of causing
an error for somebody who’s halfway through the assignment journey when
we deploy these changes. We can (and should) remove these methods and
routes maybe a week after we deploy.