Add GetServiceProviderReportingData function
The data returned from this will be used to generate a CSV for the
Service Provider to download data on all of their referrals.

This is currently missing the Referring Officer Email field, as fetching
that on the API side involves multiple calls to the Community API, which
we think will severely impact performance. We'll address that when we
can.

This also adds Cypress stubs and mock functions for this function.