Add SP CSV generator class
This uses the `csv-stringify` package to generate a CSV from a list of
ServiceProviderReportReferrals. It's also responsible for creating a
filename based on the date the report was generated.

The "Stream API"[0] is a little tricky to get your head around
initially, and took a bit of time to get into shape, but it's said to be
the most performant (which feels important with the number of referrals
we might be dealing with).

The API handily allows us to convert the CamelCase headers into
snake_case too.

[0]: https://csv.js.org/stringify/api/

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>