Display a user-facing error when a draft is not found in session
This is most likely to occur if the user tries to visit a URL that
contains a draft ID after clearing their session / session timing out.

Future improvements here could be that instead of just showing this
generic error message we show a softer, more contextual, less scary
error message, give them a link to start the journey again etc.

I got lost in the weeds of JavaScript errors when trying to write this
code - I tried subclassing Error, and a bunch of other things, and it
kept leading to errors that were missing their `stack` property etc
(vague description here because I backed away from the problem before
getting too deep into it). So in the end I worked backwards from the
errorHandler.ts code, figuring out I needed to do to make it display a
user-facing message on the error page with a 404 response, and then
copying some RestClient error generation code that would do that. I’ve
made IC-2019 to return to this and sort it out.

Also, for testing this error handling, I’ve just added an integration
test for now to sense check the whole thing. We should test this more
robustly, adding tests for errorHandler (created IC-2020 for this).