Add function for fetching Responsible Officers
from the Community API.

This is a temporary approach - we're not sure if there will be more than
one returned from the API due to the way data may be structured, so
we're planning to add a new endpoint to the Community API for returning
the current responsible officer, but just to be safe, we're returning an
array here and will display any that have `isResponsibleOfficer` set to
`true`.

As always, it's not clear whether these fields will come back in the
response from the Community API, so I've made them possibly undefined to
be safe.