Display Responsible Officer(s) on referral details
for both PPs and SPs.

The fact that we're never sure what fields will be missing from the
Community API response (they could be missing (undefined) or `null`),
the logic for what to show if nothing comes back was a bit of a pain
here, but I wanted to try and not break the user journey as much as
possible. I initially threw an error here if fields were missing, but
because have a few places where the user journey can break, I thought it
best to simply render a "not found" here.

Again, this is a temporary step, as in the future, we'll
ideally only have one Responsible Officer returned from the
endpoint with the fields we need present in the response.