Ensure we're selecting the current address for SUs
Our previous logic didn't take into account the possibility of one
address having a more recent `from` date but an elapsed `to` date,
meaning that we were displaying the incorrect address for a Service User
in some cases.

This now ensures that we're selecting the address with the most recent
`from` date that does not have a `to` date in the past.