Stop title casing intervention titles
Looks like I introduced this behaviour a while ago, but it doesn't look
right on the dashboards e.g. ETE -> Ete, East of England -> East Of
England.

I think we can rely on the case on the Intervention itself, and it's
what we're currently doing in the "Find" search results anyway.