Add contract for request to generate SP performance report
We’re rethinking the original approach that we took for service provider
data. Instead of this data being sent synchronously to the UI as JSON
and then turned into a CSV by the UI, we’ll turn it into an asychronous
process where the user submits a request for data, which is then
subsequently emailed to them by the interventions service.

The synchronous approach caused the backend team concern due to poor UX
from unbounded and unknown response times, and also a possible denial of
service vector.