Restructure appointment directories
There's some behaviour shared between PPs and SPs, and some between
initial assessment and action plan sessions, so it makes sense to group
them. It also makes things easier to visualise from the directory
structure.

The motivation behind this was originally to add in custom "behaviour"
presenters for initial assessment and action plan sessions (they
currently use a shared presenter), as we want to add custom content, and
possibly back links, to the respective pages.

Initially, the idea was to structure directories like the URLs in
`index.ts` but this has become unwieldy due to shared components.

The new structure is as follows:
```
- appointments/
 -- feedback/
  --- actionPlanSessions/
   ---- attendance/
   ---- checkYourAnswers/
   ---- confirmation/
 --- initialAssessment/
  ---- attendance/
  ---- checkYourAnswers/
  ---- confirmation/
 --- shared/
  ---- attendance/
  ---- viewFeedback/
  ---- checkYourAnswers/
  ---- confirmation/
```