Refactor Appointment models
This (large) change removes some duplication in `Appointment` data
structures, by removing the `AppointmentDetails` class that was
introduced as a shared interface that could represent both original
`Appointment` (used by for Initial Assessment)  and
`ActionPlanAppointment` classes.

There's a lot of changes in this commit due to the number of classes
that relied on these fields, so to break the important chnages, we now
have the following structure:

- `AppointmentSchedulingDetails` - scheduling details about an
appointment e.g. time, appointment type, duration. This also replaces
the `AppointmentUpdate` interface that we sent to the
Interventions Service as it contained the same fields.
- Base `Appointment` - extends `AppointmentSchedulingDetails` and adds
additional `SessionFeedback` which felt like it didn't belong on the
`SchedulingDetails` model.
- `InitialAssessmentAppointment` - extends `Appointment` and adds an
`id`.
- `ActionPlanAppointment` - extends `Appointment` and adds a
`SessionNumber`.

It's worth noting that `AppointmentSchedulingDetails` has nullable
fields, where there weren't necessarily nullable fields on the duplicate
interfaces that this new interface replaces. This shouldn't be an issue,
as we seem to have null checks for this behaviour.

This change was motivated by the need to add a `SessionType` e.g. "Group"
or "1:1" field, as I didn't want to add this to all the duplicate
classes, and now felt like a good time to re-use the shared interfaces.