Add sessionType to AppointmentSchedulingDetails
This will always be `ONE_TO_ONE` for initialAssessmentAppointment, and I
don't know if we need to display that anywhere, but setting it keeps it
simple on the UI side.

Because we use the `AppointmentSchedulingDetails` for updating an
appointment, I've had to hardcode the value to `null` so as to keep this
change relatively small and self-contained.