IC-2135: Added appointment session type to appointment table.
The types can either be ONE_TO_ONE or GROUP.

I've defaulted the existing appointments to ONE_TO_ONE but there's no way of knowing the session type of historical appointments.

For now the session type is an optional parameter; once the ui changes have gone live then I can change it to be a required parameter.