Add a check your answers page to supplier assessment scheduling journey
This converts the scheduling journey to use drafts (similarly to
be890e64), and then introduces a check your answers page in the middle
of that journey.

It maintains backwards compatibility with the existing routes. We’ll
remove these routes some time after deploying — we’re tracking this in
IC-2022.

This commit is quite large but I couldn’t see a way to split it up
without breaking the existing journey.

This changes the flow of the existing scheduling journey in the case
where the proposed schedule would cause a clash with an existing
appointment. Since we don’t find this out until we attempt to schedule
the appointment, which now doesn’t happen until after the check your
answers page, we can’t display this error inline as soon as the user
submits the scheduling details. Instead we need to redirect them back to
the scheduling form (now displaying an error) from the check your
answers page.

We need to also add a check your answers page to the action plan session
scheduling journey, but I haven’t done it here because this task already
took longer than expected. It means that there’s currently some
duplicated controller code which we’ll want to clean up later.