Added the ability to use wiremock for community-api calls.
The default url for community-api is now wiremock service.i

This was added to have control of errors being thrown in local dev run when a user does not have any active convictions when making a referral.

Wiremock has been configured to proxy any requests that start with /secure/ to community-api service, the deault proxy for all calls still goes to intervention-service but it is at a lower priority in wiremock resolutions.

Ideally I would have proxied based on the urlPattern of '/community-api.*' and '/interventions-service/.*'; however it's not possible to strip the url path when proxying (in the standalone version - it's possible to write a custom transformer if you programmatically set up wiremock service). To fix this we could extend wiremock in a new project, write the transformer and have it available as a docker container.

Alternatively I could have added a custom header only for developement env whereby we set up a pattern based on the header being present, but I didn't want to play around too much with the service classes.