Soft delete drafts when deleteDraft is called
Instead of deleting a draft from Redis using the DEL command, we instead
set the draft’s softDeleted property to true and allow Redis to clean up
the data later using its expiry mechanism.

This gives us a window of time in which we can distinguish between a
draft that we have deleted, and a draft that has expired. In real life,
this allows us to distinguish between a user pressing the Back button
after successfully completing a journey, and a user trying to interact
with an uncompleted journey after the expiry period has passed. Next
we’ll update the controller to make use of this distinction and display
relevant error messages.