Turn on searchable auditing on for our ingress
Currently, the nginx mod_security configuration is putting the audit
logs into a location that is not indexed for security reasons.

This means we _cannot_ see why a particular request is blocked, just
that it is blocked.

These changes will put a less verbose version of the audit log into our
ingress configuration, so we can find out why something got blocked, and
tailor our ruleset so we don't do false positives

Relevant discussion:
https://mojdt.slack.com/archives/C026MSZ5HRN/p1625224529031900?thread_ts=1625080484.010600&cid=C026MSZ5HRN