Display a specific message when a draft has been soft deleted
This is designed to handle the use case where the user presses the
browser’s back button after completing a journey that uses drafts (e.g.
assignment, cancellation, scheduling). Instead of showing them the page
that indicates a draft is missing (which tells them that they’ve taken
too long and have lost the answers they entered so far), we instead tell
them that the page is no longer available, and give them a link to
return to the dashboard.

The message on this page is currently generic across all journeys, but
we could make it more specific if need be.

This approach isn’t perfect – after the soft-deleted draft expires (e.g.
after 24 hours) then the user will revert to seeing the missing draft
page instead, which might cause some confusion. But hopefully this is a
sufficiently unlikely scenario.