Turn off SP performance reporting in prod.
This is because we've noticed a bug with the job where it unnecessarily tries to fetch user details from the database.
This fails if the user has not been persisted to the database, however we should still be able to generate a report based on their contracts defined via Auth.