Added two new paramters in performanceReportReader Job: 'user.authSource' and 'user.userName'.
The reason for this change is because the job was failing for some users because they didn't have any user details persisted into the database.
The job was trying to retrieve the user details from the db based on the user id passed in the job.

So instead of extracting the user from the database we instead pass all the parameters necessary to construct an AuthUser object.