Framework for pagination.
Uses the MOJ pagination component; it is quite a simple component just to display page numbers and previous/next buttons. It doesn't involve itself in the actual content for each page.
It doesn't calculate which page numbers to show, this is instead done by the new class /server/utils/pagination.ts.

The Page interface mirrors the Pageable interface within Spring.