Add routing for viewing an action plan by ID
There's a bit of refactoring involved here to extract the common logic
out of `renderActionPlan` to keep things DRY.

I've renamed the existing function for viewing an action plan to make it
explicit that this is used to render the _latest_ version of the action
plan, taken from the `Referral`'s `actionPlanId`.

This routing will be used to render previously approved action plans.