Add links to view previous Action Plans for PP
or "This version" for the action plan currently being viewed.

For linking to the latest action plan, I've opted to use the standard
`/referrals/:id/action-plan` over the new `/action-plan/:id` endpoint as
it feels strange to use the new endpoint over the one that they will
have initially used to access the action plan page from the intervention
progress page.