Fetch main address if set in Delius
The "main" address in Delius has a `status` field with `description:
"Main"` and a `code: 'M'`. which isn't the most user friendly, but
should serve our purposes. We think it's safer to rely on the `code`
field, as this seems less likely (in theory) to change than the
description.

Because we can't always rely on this field, we're falling back on the
previous logic to fetch the address with the most recent `from` date`
with a `to` date that hasn't elapsed.