IC-707/IC-1337:
Provided ability to view a single case note detail for SP or PP.

The url matches what is sent to the user by email when the case note was created.

Currently the case note can be viewed from the email - there will be a future commit which will provide a link in the view all case notes to view the single case note.