Remove nested `enabled` obj in feature flag config
Writing the feature flags in this way wasn't all that clear for both
people writing the code and reviewing it. In error, we were asserting
the existence of the `enabled` object, rather than its `enabled` value,
meaning the "Create new action plan" feature slipped through and was
live on the site before it was ready to be interacted.

This should hopefully make it clearer and more difficult for things like
this to slip through again.

I can't see any value in adding any other nested objects to the feature
object at the moment either, so I don't think we're losing out on
anything.