Preserve order of sortable tables across page visits
When a user sorts a sortable table, their chosen sort order is lost when
they next return to that page (e.g. using the browser back button). Our
user research has shown that users find this confusing, and suggested
that we try preserving the sort order across page visits. That is what
this change does.

Technically, the options for implementing this are limited. The MOJ
Frontend’s sortable table component is not designed to be interacted
with programatically, and exposes no API for either forcing a sort order
or being notified when the user changes the sort order.

I wanted to make as few assumptions as possible about the internal
implementation of the sortable table component, to make our
implementation resilient against future changes to that component. I
chose to sort the table by interacting with the table the same way that
a user would (by clicking one of the sort buttons), and to find out
about sort order changes by listening for changes to the columns’
aria-sort attributes (using a MutationObserver).