Added an update sql statement that moves appointments that exist on old action plan sessions to their equivalent session number on the new action plan where:
1. There is at least 1 appointment on the old action plan sessions.
2. There are no appointments on the new action plan sessions.
3. The number of sessions on the new action plan is enough to hold all the appointments of the old action plan.

This sql will solve an issue in prod whereby we activated multi action plan views in the front-end before the work was ready in the back-end. This meant that appointments were seemingly lost to the user; so in order to make them visible once again to the user we move them to the newest action plan.

Note that this sql will only effect around 9 referrals..