Add Referral date to My Cases table
I'd really like to be able to use one table for both SPs and PPs, but
due to the different fields required (and while we're using the
`Summary`-type component for the SP view), it's been a bit of a pain
building a one-size-fits-all approach.

We should revisit this once we've removed the `summary` component, which
we'll likely do when we introduce pagination to the table.

This commit also adds a bit more detail to the tests, which seemed to
fail when adding this new date cell.