Protect against possibly-undefined/null assessedOn
We know from past experience that fields might be missing in API
responses, or be `null` - this is mostly the case with the Community
API, but I'm adding some protections here for the `RiskSummary` response
we get back from Assess Risks and Needs to make sure this doesn't error
if there's a field missing. This may particularly be the case on the Dev
environment.