Pass AuthUser to SP intervention progress view
To access first and last name, as well as email address to be displayed.
We initially hoped to be able to use the `username` field for this, but
the `assignedTo` field doesn't include the users' real names.

This is more consistent with the PP journey, and I'm hoping we might be
able to extract common behaviour soon.

This commit has involved updating tests a fair bit for this new call to
get an AuthUser.