Add Integration tests for sorting dashboard tables
I'm hoping this is clearer and easier to add to than previous
integration tests, particularly with re-using variables and setting up a
few things in a `beforeEach` block.

We've tended to go for readability over DRYness in our integration
tests, but the amount of stubbing required to get pages to load has made
adding to the tests incredibly time-consuming and difficult to pin down
what might be missing. The newly-extracted `stubViewReferralDetails`
should hopefully take care of a lot of the difficulties we've
encountered.

I also wondered if the amount of testing here is over-the-top, but I
wanted to make sure each sort was working in the absence of any
dedicated browser-facing JS unit test suite. I opted to do it in a big
loop, again to make it easier to add to in future and to hide a lot of
the internals from readers.