Add `stubViewReferralDetails` Cypress command
This is to extract out a lot of the set up required for viewing the
"Referral Details" page on a referral into a re-useable function that
takes a requested Referral.

Adding any integration tests that visited this page required a lot of
setup, which muddied the tests' readability and made adding tests
incredibly time consuming.

This should be used when we don't really care about the information
displayed on that page, so we don't need to check any of the fields
against what's being displayed.