Changed the date for a test referral from 13th September to 13th December.
There is an issue when running these tests locally whereby shorthand September is shown as Sept but the tests expect Sep.

I assume it's to do with my local date/time setup, however I can't seem to find a way to resolve it.
Tom also said he encountered the same problem - so thought it might be wise to use a date where this issue won't occur.