User notes that they have missing sessions on their approved action plan.
The action plan was approved with 8 sessions but they can only book 5 sessions.

This is because there are two action plans tied to the referral, one with 5 and one with 8.

This was caused by a bug that was introduced into production where we deployed action plan versioning too early such that action plans can turn into an invalid state.

To fix the issue we simply just need to add the missing sessions on the action_plan_session table.