Display multiple supplier assessment appointments
for SPs.

This makes the previous functions used by the view component private,
but their behaviour is tested by the new
`supplierAssessmentAppointments` function that is now used by the view.

These aren't currently sorted in a specific order, but we'll sort them
by date.

We also want to make sure we don't show a "reschedule" button for
anything but the latest appointment that has been missed.