Sort supplier assessment appointments
These are now displayed reverse-chronolocially to the SPs.

I've added a test here for what I think is an impossible scenario,
although the compiler thinks otherwise, in which a non-current
appointment can have a `null` appointment time. This should never occur,
so I've used what I think is a fairly clear variable name in the test
(alongside a comment in the implementation code) to highlight that.