Add integration test for multiple SAA appointments
This was a struggle to get running properly due to the amount of setup
and stubbing required - I'd really like to refactor this test suite in
the near future to make a lot of the setup reusable e.g. in Cypress
commands.

However, I didn't want to spend too much time on this now, as there's
already an open PR for splitting out the Cypress tests into smaller
chunks, and I feel like refactoring can build on that.