Extended MOJFrontend.SortableTable to include second order column if specified as an attribute on the table.
If 'second-order-coloumn' is provided and it is a number and the number is a valid index of the table then this column will be used to perform a second ordering after the table has been sorted based on the chosen column.

The order will always be ascending regardless of the sort direction specified for the first column.