Remove old Action Plan session scheduling routes
These were left in when the new
`/action-plan/:id/sessions/:sessionNumber/edit/:draftBookingId/details`
routes were added to keep the deployment backwards compatible - that is,
if a user was still in the old journey and trying to schedule a
session, we didn't want to remove any of the `POST` endpoints and break
their session.

After some digging in Kibana, this should be safe to deploy now as the
last event on that route was on 28/09/2021 (7 days after these new
routes were deployed to production). There's been no traffic on the old
routes since then, only on the new ones.

This commit also updates the Cypress tests to visit the correct
endpoint, as it looks like we were still using the old ones here despite
the new ones having been added.