Add explanatory content to Add Case Note page
Users are putting anything and everything in case notes - in some cases
urgent info that should not be communicated in a case note; in other
cases irrelevant information.

Case notes are not intended to be a messaging system, but a _record_ of
stuff related to the delivery of the intervention.