Alter `auth_user` table to include new column `deleted` which acts as a boolean flag.
This flag acts as a soft deleted flag to denote if the user should no longer be used.

The flag defaults to null which means "not soft deleted".

This flag was created because there was a bug with HMPPS auth where the existing auth user ids for all users were being returned as a different value. This had the consequence of having duplicate auth users in the table for the same email address. We therefore need this flag to state which of the duplicate records are valid users or not.

We will be setting this flag to true (where applicable) in a future script.