Add new paths confirming delivery session appts
This completes the new journey for scheduling a delivery session
appointment. Again, we want to make sure the old functionality still
works as expected as links in old emails should still work until they're
no longer being used.

This isn't currently linked to in the UI, and we'll want to put it
behind a feature flag until everything is ready.