Add cypress tests for new appointment scheduling
using the new endpoints that no longer associate appointments with an
action plan.

This is mostly a copy and paste of the existing tests in
`serviceProviderReferrals.spec`, but removing the action plan-related
functionality.

This currently doesn't link to the new journey yet, as we don't want to
put it in front of users until the backend routes have been added. We'll
hide this journey behind a feature flag until the backend is ready.