Add Cypress tests for new session feedback routes
These verify that the new
`/service-provider/referral/:referralId/session/:sessionNumber/appointment/:appointmentId/...`
work as a whole journey. We're not currently linking to the new routes
yet, as we only want to do that once the backend endpoints have been
added.

There are still a couple of things in here that we'll want to change -
e.g. using the new `delivery-session/` endpoints to fetch all referrals
on the Intervention Progress page. But that will come once the new
endpoints are added. Until then, we're still stubbing requests for older
`ActionPlan` referrals here.

This also exposed an error in testing where a referral mock was being
created without an `actionPlanId`.