Some users were reporting that they were seeing that they were being assigned to the wrong referral on my cases dashboard.
After some investigation it appears that for MyCases dashboard we are returning a referral if the user has ever been assigned to it and not just if they're the latests assigend user.

This bug fix moves the custom SQL checks to be after filtering for the latest assigned referral has been done.