Create /heath/{liveness,readiness} for probes
The default `generic-service` helm chart defines `httpGet` keys that
default to health endpoints provided by Spring Boot.

When the chart configuration merges the probe configuration in our
`values.yaml` file, it **retains** the `httpGet` key, instead of
replacing it with our `tcpSocket` check.

So instead of fighting YAML merging, let's create these endpoints.

They cannot check 3rd party dependencies though, as we do not want the
service to be restarted or fail to boot if hmpps-auth or the
token-verification-service temporarily becomes unavailable.