Change the ViewUtils summaryListArgs method to properly represent email addresses when displaying caseworkers.
Changed the lines passed into the summary list args to be a special content type which the ViewUtils is responsible for displaying in the correct format.

Currently the content type could either be a string or a Caseworker object; this can be expanded in future.

This means that ViewUtils is responsible for how a caseworker should be displayed when shown as a summary list arg item.