When scheduling a delivery session appointment, if the user provides an appointmentTime that is in the past then the user is directed to a new flow where they will need to enter in session feedback details.
The full journey is using the Redis draft service so only at the final confirmation page are both scheduled appointment details and appointment feedback details submitted to the back end.

This means that the full appointment details for past appointments is sent as a single request to the backend rather than as several requests.

The existing appointment feedback flow remains untouched; only for past appointments is the draft service used.

Care has been taken to ensure that the draft for the appointment scheduling flow does not conflict with the new draft which include feedback details.
In order to ensure that old drafts can work, we've appended feedback details as a new undefinable field in the existing draft so it should be compatible with the old and new draft.