Set the sessionFeedback to be "undefineable" for a DraftAppointment and remove setting empty session feedback details until the POST feedback attendance.
The main reason behind this is to prevent a case where the code is deployed whilst the user has defined a past appointment and they're viewing the check your answers page and about to press submit. We need to ensure that the GET/POST feedback attendance with draftId doesnt break.